Moved embeds to own utils file
